+++ wc_3dv_panel.py
@@ -1,0 +1,390 @@
+
+import bpy, os, re
+from bpy.props import StringProperty, EnumProperty, BoolProperty, IntProperty
+from bpy.types import Panel, Operator, Menu
+
+
+def engine_compatibility(context):
+    scn = context.scene
+    return scn.render.engine == 'CYCLES'
+
+def find_basic(nt):
+    return nt.nodes.find("WorldControlBasic")
+
+def find_advanced(nt):
+    return nt.nodes.find("WorldControlAdvanced")
+    
+def check_world_control_type(context):
+    world = bpy.context.scene.world
+    nt = world.node_tree
+#    print(find_basic(nt) != -1)
+    if find_basic(nt) != -1:
+        return "World Control Basic"
+    if find_advanced(nt) != -1:
+        return "World Control Advanced"
+    else:
+        return "World Control"
+
+def return_world_control_type():
+    world = bpy.context.scene.world
+    nt = world.node_tree
+#    print(find_basic(nt) != -1)
+    if find_basic(nt) != -1:
+        return "WorldControlBasic"
+    if find_advanced(nt) != -1:
+        return "WorldControlAdvanced"
+    else:
+        return "World Control"
+
+
+class WorldControlPanel:
+    bl_space_type = "VIEW_3D"
+    bl_region_type = "UI"
+    bl_category = 'View'
+
+
+class WC_PT_main_panel(WorldControlPanel, Panel):          
+    bl_label = ""
+    #bl_context = "objectmode"
+    
+    @classmethod
+    def poll(cls, context):
+        world = bpy.context.scene.world
+        nt = world.node_tree
+        basic = find_basic(nt)
+        advanced = find_advanced(nt)
+#        print(engine_compatibility(context))
+#        print(basic or advanced != -1)
+        return (basic != -1 or advanced != -1) and engine_compatibility(context)
+    
+    def draw_header(self, context):
+        layout = self.layout
+        layout.label(text=str(check_world_control_type(context)))
+        
+    def draw(self, context):
+        scn = context.scene
+        layout = self.layout
+        layout.use_property_split = True
+        layout.use_property_decorate = False
+                
+        col = layout.column()
+        
+        world = bpy.context.scene.world
+        nodes = scn.world.node_tree.nodes
+        nt = world.node_tree
+        basic = find_basic(nt)
+        advanced = find_advanced(nt)
+       
+        if engine_compatibility and (basic != -1 or advanced != -1):
+            reset_bg = col.operator("wc.reset_settings", icon = 'LOOP_BACK')
+            reset_bg.reset_type = 'BG'         
+                                            
+        if (basic == -1 and advanced == -1):   
+            col.label(text = 'No World Control in shader', icon = 'INFO') 
+        if not engine_compatibility:
+            col.label(text = 'Not compatible with this render engine', icon = 'INFO')
+            col.prop(scn.render, 'engine')
+#            col.separator()
+        else:
+            col.prop(nodes['Mapping'].inputs[2], "default_value", text = 'Rotation')   
+            
+class WC_PT_light_settings(WorldControlPanel, Panel):
+    bl_category = "View"
+    # bl_context = ".objectmode"  # dot on purpose (access from topbar)
+    bl_label = "Light"
+    bl_parent_id = "WC_PT_main_panel"
+    # bl_options = {'DEFAULT_CLOSED'}
+    
+    @classmethod
+    def poll(cls, context):
+        world = bpy.context.scene.world
+        nt = world.node_tree
+        basic = find_basic(nt)
+        advanced = find_advanced(nt)
+#        print(basic or advanced != -1)
+        return (basic != -1 or advanced != -1) and engine_compatibility(context)
+
+    def draw(self, context):
+        layout = self.layout
+
+        layout.use_property_split = True
+        layout.use_property_decorate = False
+        scn = context.scene
+                 
+        col = layout.column(align=True)                
+        nodes = scn.world.node_tree.nodes
+        controlType = return_world_control_type()  
+        
+#        col.prop(nodes['Mapping'].inputs[2], "default_value", text = 'Rotation')   
+             
+        if controlType == 'WorldControlBasic':
+            col.prop(nodes['WorldControlBasic'].inputs[1], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlBasic'].inputs[2], "default_value", text = 'Temperature')
+            
+            
+        if controlType == 'WorldControlAdvanced':
+            col.prop(nodes['WorldControlAdvanced'].inputs[1], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlAdvanced'].inputs[2], "default_value", text = 'Saturation')
+            col.prop(nodes['WorldControlAdvanced'].inputs[3], "default_value", text = 'Temperature')
+
+
+class WC_PT_background_settings(WorldControlPanel, Panel):
+    bl_category = "View"
+    # bl_context = ".objectmode"  # dot on purpose (access from topbar)
+    bl_label = "Background"
+    bl_parent_id = "WC_PT_main_panel"
+    bl_options = {'DEFAULT_CLOSED'}
+    
+    @classmethod
+    def poll(cls, context):
+        world = bpy.context.scene.world
+        nt = world.node_tree
+        basic = find_basic(nt)
+        advanced = find_advanced(nt)
+        return (basic != -1 or advanced != -1) and engine_compatibility(context)
+
+    def draw(self, context):
+        layout = self.layout
+
+        layout.use_property_split = True
+        layout.use_property_decorate = False
+        scn = context.scene
+
+        col = layout.column(align=True)                
+        nodes = scn.world.node_tree.nodes
+        
+        controlType = return_world_control_type()  
+     
+        if controlType == 'WorldControlBasic':
+            col.prop(nodes['WorldControlBasic'].inputs[3], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlBasic'].inputs[4], "default_value", text = 'Saturation')
+            
+        if controlType == 'WorldControlAdvanced':
+            col.prop(nodes['WorldControlAdvanced'].inputs[7], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlAdvanced'].inputs[8], "default_value", text = 'Saturation') 
+
+
+class WC_PT_diffuse_settings(WorldControlPanel, Panel):
+    bl_category = "View"
+    # bl_context = ".objectmode"  # dot on purpose (access from topbar)
+    bl_label = "Diffuse"
+    bl_parent_id = "WC_PT_main_panel"
+    bl_options = {'DEFAULT_CLOSED'}
+    
+    @classmethod
+    def poll(cls, context):
+        world = bpy.context.scene.world
+        nt = world.node_tree
+        basic = find_basic(nt)
+        advanced = find_advanced(nt)
+        return (basic != -1 or advanced != -1) and engine_compatibility(context)
+
+    def draw(self, context):
+        layout = self.layout
+
+        layout.use_property_split = True
+        layout.use_property_decorate = False
+        scn = context.scene
+
+        col = layout.column(align=True)                
+        nodes = scn.world.node_tree.nodes
+        
+        controlType = return_world_control_type()  
+     
+        if controlType == 'WorldControlBasic':
+            col.prop(nodes['WorldControlBasic'].inputs[5], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlBasic'].inputs[6], "default_value", text = 'Saturation')      
+            
+        if controlType == 'WorldControlAdvanced':
+            col.prop(nodes['WorldControlAdvanced'].inputs[4], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlAdvanced'].inputs[5], "default_value", text = 'Saturation')                         
+
+
+class WC_PT_reflections_settings(WorldControlPanel, Panel):
+    bl_category = "View"
+    # bl_context = ".objectmode"  # dot on purpose (access from topbar)
+    bl_label = "Reflections"
+    bl_parent_id = "WC_PT_main_panel"
+    bl_options = {'DEFAULT_CLOSED'}
+    
+    @classmethod
+    def poll(cls, context):
+        world = bpy.context.scene.world
+        nt = world.node_tree
+        basic = find_basic(nt)
+        advanced = find_advanced(nt)
+        return (advanced != -1) and engine_compatibility(context)
+
+    def draw(self, context):
+        layout = self.layout
+
+        layout.use_property_split = True
+        layout.use_property_decorate = False
+        scn = context.scene
+
+        col = layout.column(align=True)                
+        nodes = scn.world.node_tree.nodes
+        
+        controlType = return_world_control_type()  
+     
+        if controlType == 'WorldControlAdvanced':
+            col.prop(nodes['WorldControlAdvanced'].inputs[9], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlAdvanced'].inputs[10], "default_value", text = 'Saturation') 
+
+
+class WC_PT_glossy_settings(WorldControlPanel, Panel):
+    bl_category = "View"
+    # bl_context = ".objectmode"  # dot on purpose (access from topbar)
+    bl_label = "Glossy"
+    bl_parent_id = "WC_PT_main_panel"
+    bl_options = {'DEFAULT_CLOSED'}
+    
+    @classmethod
+    def poll(cls, context):
+        world = bpy.context.scene.world
+        nt = world.node_tree
+        basic = find_basic(nt)
+        advanced = find_advanced(nt)
+        return (basic != -1) and engine_compatibility(context)
+
+    def draw(self, context):
+        layout = self.layout
+
+        layout.use_property_split = True
+        layout.use_property_decorate = False
+        scn = context.scene
+
+        col = layout.column(align=True)                
+        nodes = scn.world.node_tree.nodes
+        
+        controlType = return_world_control_type()  
+     
+        if controlType == 'WorldControlBasic':
+            col.prop(nodes['WorldControlBasic'].inputs[7], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlBasic'].inputs[8], "default_value", text = 'Saturation')   
+
+
+class WC_PT_refraction_settings(WorldControlPanel, Panel):
+    bl_category = "View"
+    # bl_context = ".objectmode"  # dot on purpose (access from topbar)
+    bl_label = "Refraction"
+    bl_parent_id = "WC_PT_main_panel"
+    bl_options = {'DEFAULT_CLOSED'}
+    
+    @classmethod
+    def poll(cls, context):
+        world = bpy.context.scene.world
+        nt = world.node_tree
+        basic = find_basic(nt)
+        advanced = find_advanced(nt)
+        return (advanced != -1) and engine_compatibility(context)
+
+    def draw(self, context):
+        layout = self.layout
+
+        layout.use_property_split = True
+        layout.use_property_decorate = False
+        scn = context.scene
+
+        col = layout.column(align=True)                
+        nodes = scn.world.node_tree.nodes
+        
+        controlType = return_world_control_type()  
+        
+        if controlType == 'WorldControlAdvanced':
+            col.prop(nodes['WorldControlAdvanced'].inputs[11], "default_value", text = 'Strength')
+            col.prop(nodes['WorldControlAdvanced'].inputs[12], "default_value", text = 'Saturation')            
+
+
+
+def reset_settings(controlType):
+    scn = bpy.context.scene
+    worlds = bpy.data.worlds
+    world = scn.world
+    nodes = world.node_tree.nodes
+    
+    # Light
+    if controlType == 'WorldControlBasic':
+        nodes['WorldControlBasic'].inputs[1].default_value = 1.0 # Strength
+        nodes['WorldControlBasic'].inputs[2].default_value = 5500 # Temperature
+
+    if controlType == 'WorldControlAdvanced':
+        nodes['WorldControlAdvanced'].inputs[1].default_value = 1.0 # Strength
+        nodes['WorldControlAdvanced'].inputs[2].default_value = 0.5 # Saturation
+        nodes['WorldControlAdvanced'].inputs[3].default_value = 5500 # Temperature
+
+    # Background
+    if controlType == 'WorldControlBasic':
+        nodes['WorldControlBasic'].inputs[3].default_value = 1.0 # Strength
+        nodes['WorldControlBasic'].inputs[4].default_value = 0.5 # Saturation
+        
+    if controlType == 'WorldControlAdvanced':
+        nodes['WorldControlAdvanced'].inputs[7].default_value = 1.0 # Strength
+        nodes['WorldControlAdvanced'].inputs[8].default_value = 0.5 # Saturation
+
+    # Diffuse
+    if controlType == 'WorldControlBasic':
+        nodes['WorldControlBasic'].inputs[5].default_value = 1.0 # Strength
+        nodes['WorldControlBasic'].inputs[6].default_value = 0.5 # Saturation
+        
+    if controlType == 'WorldControlAdvanced':
+        nodes['WorldControlAdvanced'].inputs[4].default_value = 1.0 # Strength
+        nodes['WorldControlAdvanced'].inputs[5].default_value = 0.5 # Saturation
+
+    # Reflections
+    if controlType == 'WorldControlAdvanced':
+        nodes['WorldControlAdvanced'].inputs[9].default_value = 1.0 # Strength
+        nodes['WorldControlAdvanced'].inputs[10].default_value = 0.5 # Saturation
+    # Glossy
+    if controlType == 'WorldControlBasic':
+        nodes['WorldControlBasic'].inputs[7].default_value = 1.0 # Strength
+        nodes['WorldControlBasic'].inputs[8].default_value = 0.5 # Saturation
+
+    # Refraction
+    if controlType == 'WorldControlAdvanced':
+        nodes['WorldControlAdvanced'].inputs[11].default_value = 1.0 # Strength
+        nodes['WorldControlAdvanced'].inputs[12].default_value = 0.5 # Saturation
+
+
+class WC_OT_reset_settings(Operator):
+    bl_idname = "wc.reset_settings"
+    bl_label = "Reset"
+    bl_description = "Reset to Default Values"
+    bl_options = {'REGISTER', 'UNDO'}
+    
+    controlType : StringProperty(default = 'WorldControlBasic')
+    
+    # @classmethod
+    # def poll(cls, context):
+    #     return check_world_nodes() == 'OK'
+    
+    def execute(self, context):
+        controlType = return_world_control_type()
+        reset_settings(controlType)
+        return {'FINISHED'}
+
+classes = (
+    WC_PT_main_panel,
+    WC_PT_light_settings,
+    WC_PT_background_settings,
+    WC_PT_diffuse_settings,
+    WC_PT_reflections_settings,
+    WC_PT_glossy_settings,
+    WC_PT_refraction_settings,
+    WC_OT_reset_settings
+    )
+    
+def register():
+    from bpy.utils import register_class
+    for cls in classes:
+        register_class(cls)        
+
+def unregister():
+    from bpy.utils import unregister_class
+    for cls in reversed(classes):
+        unregister_class(cls)        
+
+
+
+if __name__ == "__main__":
+    register()