+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2021-29-10
+### Added
+- Reset operator > set all inputs to default
+- Panel view easy access from 3d viewport
+
 ## [0.0.2] - 2021-28-10
 ### Fixed 
 - The Add function was visible in materials nodes as well, it added empty node in material, added poll to customnodecategrory 
@@ -18,6 +23,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
-
+[0.0.3]:https://github.com/schroef/World_Control/releases/tag/v0.0.3
 [0.0.2]:https://github.com/schroef/World_Control/releases/tag/v0.0.2
 [0.0.1]:https://github.com/schroef/World_Control/releases/tag/v0.0.1